Refactor credential renewal into Java client
